Create md5check2.sh
Checks to see if hashes are the same or there has been changes. The command is `md5sum -c filename.iso.md5`. It displays the output on the console, and saves the output to a file.  Then it checks the exit status array to see if no changes has been made (0) or there has been changes (1).